Create PendingIntentFactory
