+++ libraries/push/impl/src/main/kotlin/io/element/android/libraries/push/impl/notifications/factories/action/RejectInvitationActionFactory.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.libraries.push.impl.notifications.factories
+package io.element.android.libraries.push.impl.notifications.factories.action
 
 import android.app.PendingIntent
 import android.content.Context
@@ -30,19 +30,18 @@ import io.element.android.services.toolbox.api.strings.StringProvider
 import io.element.android.services.toolbox.api.systemclock.SystemClock
 import javax.inject.Inject
 
-class AcceptInvitationActionFactory @Inject constructor(
+class RejectInvitationActionFactory @Inject constructor(
     @ApplicationContext private val context: Context,
     private val actionIds: NotificationActionIds,
     private val stringProvider: StringProvider,
     private val clock: SystemClock,
 ) {
-    // offer to type a quick accept button
-    fun create(inviteNotifiableEvent: InviteNotifiableEvent): NotificationCompat.Action {
+    fun create(inviteNotifiableEvent: InviteNotifiableEvent): NotificationCompat.Action? {
         val sessionId = inviteNotifiableEvent.sessionId
         val roomId = inviteNotifiableEvent.roomId
         val intent = Intent(context, NotificationBroadcastReceiver::class.java)
-        intent.action = actionIds.join
-        intent.data = createIgnoredUri("acceptInvite?${sessionId.value}&${roomId.value}")
+        intent.action = actionIds.reject
+        intent.data = createIgnoredUri("rejectInvite?${sessionId.value}&${roomId.value}")
         intent.putExtra(NotificationBroadcastReceiver.KEY_SESSION_ID, sessionId)
         intent.putExtra(NotificationBroadcastReceiver.KEY_ROOM_ID, roomId)
         val pendingIntent = PendingIntent.getBroadcast(
@@ -51,9 +50,10 @@ class AcceptInvitationActionFactory @Inject constructor(
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
+
         return NotificationCompat.Action.Builder(
-            R.drawable.vector_notification_accept_invitation,
-            stringProvider.getString(R.string.notification_invitation_action_join),
+            R.drawable.vector_notification_reject_invitation,
+            stringProvider.getString(R.string.notification_invitation_action_reject),
             pendingIntent
         ).build()
     }
